Add a guide on how to self-host the OpenFunctions model (#157)
This PR introduces a guide for users to self-host the OpenFunctions
model. It includes detailed instructions and example code for setting up
a local server using FastAPI and uvicorn, and for configuring a client
to interact with this server using the OpenAI package. The server mimics
the openai.ChatCompletion.create API interface, ensuring compatibility
with OpenAI's request structure, while the client demonstrates
interaction with the self-hosted model, akin to using Gorilla hosted
servers.

This addition aims to enhance user understanding and ease the process of
local deployment and usage of the OpenFunctions model.

Resolves #146

---------

Co-authored-by: Raman Varma <ramanvarma@Ramans-MacBook-Pro.local>