[BFCL] Add Test Dataset to Repository (#504)
Previously, the test dataset was stored on HuggingFace, requiring users
to clone this repository, download the dataset separately from
HuggingFace, and then run the evaluation pipeline. This has caused many
inconveniences and confusion within the community. Users often struggled
with the inconsistency of having the possible answers within the
repository while the test dataset was missing.

Partially addresses #501.