[BFCL] Set Model Temperature to 0.001 for All Models (#574)
The current model response generation script uses a default temperature
of 0.7 for inference. This introduces some degree of randomness into the
model output generation, leading to potential variability in the
evaluation scores from run to run.
For benchmarking purposes, we set it to 0.001 for consistency and
reliability of the evaluation results.

resolves #500 , resolves #562 

This will affect the leaderboard score. We will update it shortly.

---------

Co-authored-by: Shishir Patil <30296397+ShishirPatil@users.noreply.github.com>