[BFCL Chore] Ensure Correct Input Format for Eval Checker (#860)
In some cases, the model handler’s decode_ast method returns
successfully but produces output in an unexpected format, causing issues
in downstream evaluations that do not perform argument format
validation. This problem is especially common when the model does not
output any function calls, resulting in a human-readable string instead
of the expected structure.

This PR refines the `is_function_calling_format_output` function to
enforce that outputs must be a list of dictionaries in the following
format before calling the checker function:
```
[
  {func1: {param1: val1, param2: val2, ...}},
  {func2: {param1: val1, param2: val2, ...}},
  ...
]
```

Note: This PR will not affect the leaderboard score.