[BFCL] Support Local Inference for `deepseek-ai/DeepSeek-R1` (#926)
This PR builds on #901 to add support for the `DeepSeek-R1` models by
enabling local inference. While #901 introduced inference using the
official DeepSeek API endpoint, it is often unstable and overloaded,
making it impractical to run the full BFCL benchmarks at the moment
(which could take days via API alone). Local inference provides a more
efficient alternative for evaluating `DeepSeek-R1` models.

You can use model names to differentiate between the two inference
methods:

- `deepseek-ai/DeepSeek-R1` for local inference
- `DeepSeek-R1` for API inference

Note that #901 is a prerequisite for this PR and should be merged first,
as it includes the necessary modifications to the base handler that this
PR also depends on.